Ensure constants and utils are under common
