+++ package.json
@@ -3,6 +3,10 @@
   "version": "2.19.0.0",
   "description": "OpenSearch Dashboards plugin for Query Insights",
   "main": "index.js",
+  "opensearchDashboards": {
+    "version": "2.19.0",
+    "templateVersion": "2.19.0"
+  },
   "license": "Apache-2.0",
   "homepage": "https://github.com/opensearch-project/query-insights-dashboards",
   "config": {
