+++ public/pages/QueryDetails/Components/QuerySummary.tsx
@@ -16,7 +16,7 @@ import {
   TIMESTAMP,
   TOTAL_SHARDS,
 } from '../../../../common/constants';
-import { calculateMetric } from '../../Utils/MetricUtils';
+import { calculateMetric } from '../../../../common/utils/MetricUtils';
 
 // Panel component for displaying query detail values
 const PanelItem = ({ label, value }: { label: string; value: string | number }) => (
