+++ public/pages/QueryGroupDetails/QueryGroupDetails.tsx
@@ -29,7 +29,7 @@ import { QueryGroupSampleQuerySummary } from './Components/QueryGroupSampleQuery
 import { QueryInsightsDashboardsPluginStartDependencies } from '../../types';
 import { PageHeader } from '../../components/PageHeader';
 import { SearchQueryRecord } from '../../../types/types';
-import { retrieveQueryById } from '../Utils/QueryUtils';
+import { retrieveQueryById } from '../../../common/utils/QueryUtils';
 import { QueryInsightsDataSourceMenu } from '../../components/DataSourcePicker';
 import { getDataSourceFromUrl } from '../../utils/datasource-utils';
